chore: moved tmux logic to own file
